+++ src/background/widgets/webview.rs
@@ -8,12 +8,14 @@ use seelen_core::{
 };
 
 use tauri::Manager;
+use windows::Win32::Foundation::HWND;
 
 use crate::{
     app::get_app_handle,
     error::{Result, ResultLogExt},
     state::application::FULL_STATE,
     utils::constants::SEELEN_COMMON,
+    windows_api::WindowsApi,
 };
 
 pub struct WidgetWebview(pub tauri::WebviewWindow);
@@ -100,6 +102,9 @@ impl WidgetWebview {
             .additional_browser_args(&args.to_string())
             .build()?;
 
+        // Widgets handle their own show/hide animations, avoid the ones from the system.
+        // SAFETY: HWND in windows 0.61 (tauri) and 0.62 (ours) share the same memory layout
+        WindowsApi::set_system_transitions_disabled(HWND(window.hwnd()?.0), true).log_error();
         Ok(Self(window))
     }